upd(Instagram): Code correction for `Selectable bio` patch
