--- src/main/kotlin/app/revanced/patches/instagram/misc/selectableBio/fingerprints/SelectableBioFingerprint.kt
@@ -1,8 +1,0 @@
-package app.revanced.patches.instagram.misc.selectableBio.fingerprints
-
-import app.revanced.patcher.fingerprint.MethodFingerprint
-
-internal object SelectableBioFingerprint:MethodFingerprint(
-    strings = listOf("is_bio_visible"),
-    returnType = "V"
-)
