+++ src/main/kotlin/app/revanced/patches/instagram/patches/selectableBio/fingerprints/SelectableBioFingerprint.kt
@@ -1,0 +1,8 @@
+package app.revanced.patches.instagram.patches.selectableBio.fingerprints
+
+import app.revanced.patcher.fingerprint.MethodFingerprint
+
+internal object SelectableBioFingerprint : MethodFingerprint(
+    strings = listOf("is_bio_visible"),
+    returnType = "V",
+)
